Move ChannelSet extensions to :core:model
