+++ app/src/main/java/com/geeksville/mesh/model/UIState.kt
@@ -78,7 +78,9 @@ import org.meshtastic.core.database.model.Node
 import org.meshtastic.core.datastore.UiPreferencesDataSource
 import org.meshtastic.core.model.DataPacket
 import org.meshtastic.core.model.DeviceHardware
+import org.meshtastic.core.model.util.getChannel
 import org.meshtastic.core.model.util.getShortDate
+import org.meshtastic.core.model.util.toChannelSet
 import org.meshtastic.core.strings.R
 import javax.inject.Inject
